+++ truss/tests/test_model_inference.py
@@ -194,29 +194,6 @@ def test_requirements_file_truss():
 
 
 @pytest.mark.integration
-def test_docker_server_truss():
-    with ensure_kill_all():
-        truss_root = Path(__file__).parent.parent.parent.resolve() / "truss"
-
-        truss_dir = truss_root / "test_data" / "test_docker_server_truss"
-
-        tr = TrussHandle(truss_dir)
-        LocalConfigHandler.set_secret("hf_access_token", "123")
-        _ = tr.docker_run(local_port=8090, detach=True, wait_for_server_ready=True)
-        truss_server_addr = "http://localhost:8090"
-        full_url = f"{truss_server_addr}/v1/models/model:predict"
-
-        response = requests.post(full_url, json={})
-        assert response.status_code == 200
-        assert response.json() == {
-            "message": "Hello World",
-            "is_torch_cuda_available": False,
-            "is_env_var_passed": True,
-            "is_secret_mounted": True,
-        }
-
-
-@pytest.mark.integration
 @pytest.mark.parametrize("pydantic_major_version", ["1", "2"])
 def test_requirements_pydantic(pydantic_major_version):
     with ensure_kill_all():
